101-print_number.c

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -1,4 +1,5 @@
 #include "main.h"
+#include <stdio.h>
 
 /**
  * print_number - print numbers chars
@@ -22,5 +23,5 @@ void print_number(int n)
 	{
 		print_number(n1 / 10);
 	}
-	_putchar((n1 % 10) + '0');
+	_putchar((unsigned int) n1 % 10 + '0');
 }
